requirements.txt
google-api-python-client==2.107.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.7.0
anthropic
openai>=1.0.0
mysql-connector-python
pandas
sqlalchemy
pymysql 